+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `preprocess()` now ensures that the detected contact point has the lowest height.
 
+### Deprecated
+
+- `RawProfileBase.all_profiles()` method is deprecated. Directly iterate `profiles()` method instead.
+
 ## [1.5.0] - 2025-10-21
 
 ### Added